zha support for ICTCG1Controller
This adds zha support for the ICTC-G-1 controller.
The device is VERY chatty with zha_event so the sometimes the behaviour is a little wonky. The dimmer I tested with isn't the latest firmware so I will try and update it and see if that makes any difference.
Note that I used Light.HOLD_BRIGHTNESS_DOWN and Light.HOLD_BRIGHTNESS_UP twice as there are two different args, depending on the rate at which you rotate the dimmer - slow or fast. This is still separate from "quick" right and left. I tried with no args (just "move" and "move_with_on_off" but then I got no dimming at all with rotation - maybe an issue with args processing?