fix(light_controller): Change default value for add_transition_turn_toggle from True to False
BREAKING CHANGE: There is less people using transition attribute for on/off calls than people having problems with this and having to explicitly add the add_transition_turn_toggle: false to all ControllerX configurations. I first thought that this will get fix for the third parties soon, but seeing that this still happens for Hue and deCONZ, I considered this.