refactor(file-names): change file names to something more specific by adding the "cx" (controllerx) as prefix
each AppDaemon app does not have its own environment, so all modules are shared. This means that if someone else has ControllerX installed and wants to create a module named "conts.py" as I had, then there will be conflicts due to sharing the same environment. For this reason, I added a custom prefix to all modules to lower the probability of someone running to this problem.